fix failed to save challenge error
The error was caused by an incorrect array index access in the
saveChallengeDraftAction. The code was trying to access companyResult[1]
(the second element) instead of companyResult[0] (the first element)
when retrieving the company data from the database query result.

What was fixed:
Changed const company = companyResult[1]; to const company =
companyResult[0]; on line 42 of save-challenge-draft.action.ts

Why this happened:
Database queries return arrays of results
When you expect a single result, you should access the first element
([0])
The code was incorrectly accessing the second element ([1]), which was
undefined
This caused the error when trying to access company.id